Split workspace surface parity gates
